+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.12.26+1
 
 * Fix lower bound on package `stack_trace`. Now 1.6.0.
+* Manually close browser process streams to prevent test hangs.
 
 ## 0.12.26